Split plugin into pylons/flask mixins
